feat: Add get_absolute_url method to the Post model
This commit adds a `get_absolute_url` method to the `Post` model in the `blog` app. The `get_absolute_url` method returns the URL for the detail view of a blog post, using the `reverse` function from Django's `urls` module. This method will be used to generate the URL for each blog post in the post list template.